fix: cleanup log event component and split into smaller chunks
