+++ clients/web/src/lib/console/process-log-event-for-view.ts
@@ -5,6 +5,10 @@ import { useMonitor } from "~/context/monitor-provider";
 import { processFieldValue } from "~/lib/span/process-field-value";
 import { LogEvent } from "../proto/logs";
 
+export type ProcessedLogEvent = NonNullable<
+  ReturnType<typeof processLogEventForView>
+>;
+
 export function processLogEventForView(logEvent: LogEvent) {
   const { monitorData } = useMonitor();
   const { message, at, fields, parent } = logEvent;
